chore: organize comlink entities
